+++ ChangeLog.md
@@ -43,6 +43,9 @@ See docs/process.md for more on how version tagging works.
   a Wasm Worker.  This mode increases the memory used by each Wasm Worker by
   ~500 bytes (in the same way that declaring ~500 bytes of TLS data would).
   (#26757)
+- The `EMSCRIPTEN_DEVICE_MOTION_EVENT_SUPPORTS_XXX` macros are now deprecated
+  since all device motion event fields are always supported on modern browsers.
+
 
 5.0.6 - 04/14/26
 ----------------